#5412: moved PublisherConfig to cdp package
